For some reason chrome does not support background.type in manifest.json. Doing a workaround.
