+++ package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "devDependencies": {
     "eslint": "^8.39.0",
     "eslint-config-airbnb": "^19.0.4",
